Enviando aulas

+++ app_locadora_carros/app/Http/Controllers/MarcaController.php
@@ -38,6 +38,11 @@ class MarcaController extends Controller
     public function show($id)
     {
         $marca = $this->marca->find($id);
+        if($marca === null) {
+            return [
+                'erro' => 'Recurso pesquisado não existe'
+            ];
+        }
         return $marca;
     }
 
@@ -49,6 +54,11 @@ class MarcaController extends Controller
     {
         // $marca->update($request->all());
         $marca = $this->marca->find($id);
+        if(!$marca) {
+            return [
+                'erro' => 'Impossível realizar a atualização. O recurso solicitado não existe'
+            ];
+        }
         $marca->update($request->all());
         return $marca;
     }
@@ -60,6 +70,11 @@ class MarcaController extends Controller
     public function destroy($id)
     {
         $marca = $this->marca->find($id);
+        if(!$marca) {
+            return [
+                'erro' => 'Impossível realizar a exclusão. O recurso solicitado não existe'
+            ];
+        }
         $marca->delete();
         return ['msg' => 'A marca foi removida com sucesso!'];
     }
